requirements.txt
Flask
mysql-connector-python
bcrypt
gunicorn
scikit-learn
flask-cors

